add benchmark test to ci
