+++ Makefile.common
@@ -66,6 +66,11 @@ test:
 	@echo ">> running all tests"
 	$(GO) test -race $(pkgs)
 
+.PHONY: bench
+bench:
+	@echo ">> running all benchmarks"
+	$(GO) test -bench . -race $(pkgs) 
+
 .PHONY: format
 format:
 	@echo ">> formatting code"
